@@ -11,6 +11,7 @@
         "ext-intl": "*",
         "ext-json": "*",
         "ext-soap": "*",
+        "ext-zip": "*",
         "api-platform/core": "^2.6",
         "composer/package-versions-deprecated": "^1.8",
         "craue/twigextensions-bundle": "^2.4",
@@ -41,32 +42,33 @@
         "sabre/vobject": "^4.3",
         "sensio/framework-extra-bundle": "^5.6",
         "sylius/theme-bundle": "^2.2",
-        "symfony/asset": "4.4.*",
-        "symfony/console": "4.4.*",
-        "symfony/dotenv": "4.4.*",
-        "symfony/expression-language": "4.4.*",
+        "symfony/asset": "5.4.*",
+        "symfony/console": "5.4.*",
+        "symfony/dotenv": "5.4.*",
+        "symfony/expression-language": "5.4.*",
         "symfony/flex": "^1.3",
-        "symfony/form": "4.4.*",
-        "symfony/framework-bundle": "4.4.*",
-        "symfony/http-client": "4.4.*",
-        "symfony/ldap": "4.4.*",
-        "symfony/lock": "4.4.*",
-        "symfony/mailer": "4.4.*",
-        "symfony/messenger": "4.4.*",
+        "symfony/form": "5.4.*",
+        "symfony/framework-bundle": "5.4.*",
+        "symfony/http-client": "5.4.*",
+        "symfony/ldap": "5.4.*",
+        "symfony/lock": "5.4.*",
+        "symfony/mailer": "5.4.*",
+        "symfony/messenger": "5.4.*",
         "symfony/monolog-bundle": "^3.3",
         "symfony/polyfill-apcu": "^1.10",
-        "symfony/process": "4.4.*",
-        "symfony/property-access": "4.4.*",
-        "symfony/property-info": "4.4.*",
-        "symfony/proxy-manager-bridge": "4.4.*",
-        "symfony/security-bundle": "4.4.*",
-        "symfony/serializer": "4.4.*",
-        "symfony/stopwatch": "4.4.*",
-        "symfony/translation": "4.4.*",
-        "symfony/twig-bundle": "4.4.*",
-        "symfony/validator": "4.4.*",
+        "symfony/process": "5.4.*",
+        "symfony/property-access": "5.4.*",
+        "symfony/property-info": "5.4.*",
+        "symfony/proxy-manager-bridge": "5.4.*",
+        "symfony/security-bundle": "5.4.*",
+        "symfony/serializer": "5.4.*",
+        "symfony/stopwatch": "5.4.*",
+        "symfony/translation": "5.4.*",
+        "symfony/twig-bundle": "5.4.*",
+        "symfony/validator": "5.4.*",
         "symfony/webpack-encore-bundle": "^1.6",
-        "symfony/yaml": "4.4.*",
+        "symfony/workflow": "5.4.*",
+        "symfony/yaml": "5.4.*",
         "tomnomnom/etherpad-lite-client": "^0.1.2",
         "twig/extra-bundle": "^3.3",
         "twig/string-extra": "^3.3",
@@ -78,6 +80,7 @@
         "codeception/module-datafactory": "^2.0",
         "codeception/module-db": "^2.0",
         "codeception/module-doctrine2": "^2.0",
+        "codeception/module-filesystem": "^2.0",
         "codeception/module-phpbrowser": "^2.0",
         "codeception/module-rest": "^2.0",
         "codeception/module-symfony": "^2.1",
@@ -85,7 +88,7 @@
         "doctrine/doctrine-fixtures-bundle": "^3.3",
         "phpcompatibility/php-compatibility": "^9.3",
         "symfony/maker-bundle": "^1.21",
-        "symfony/web-profiler-bundle": "4.4.*"
+        "symfony/web-profiler-bundle": "5.4.*"
     },
     "config": {
         "optimize-autoloader": true,
@@ -139,7 +142,7 @@
     "extra": {
         "symfony": {
             "allow-contrib": true,
-            "require": "4.4.*"
+            "require": "5.4.*"
         },
         "ckeditor-tag": "4.11.4",
         "ckeditor-clear": "drop"
