@@ -1,14 +1,13 @@
 <?php
 namespace App\Filter;
 
+use App\Entity\Labels;
+use App\Repository\LabelRepository;
+use App\Utils\RoomService;
+use Lexik\Bundle\FormFilterBundle\Filter\Form\Type as Filters;
 use Symfony\Component\Form\AbstractType;
 use Symfony\Component\Form\FormBuilderInterface;
 use Symfony\Component\HttpFoundation\RequestStack;
-
-use App\Utils\RoomService;
-use App\Repository\LabelRepository;
-
-use Lexik\Bundle\FormFilterBundle\Filter\Form\Type as Filters;
 
 class RubricFilterType extends AbstractType
 {
@@ -25,7 +24,7 @@ class RubricFilterType extends AbstractType
      * Builds the form.
      * This method is called for each type in the hierarchy starting from the top most type.
      * Type extensions can further modify the form.
-     * 
+     *
      * @param  FormBuilderInterface $builder The form builder
      * @param  array                $options The options
      */
@@ -45,7 +44,7 @@ class RubricFilterType extends AbstractType
                     $builder
                         ->add('group', Filters\EntityFilterType::class, array(
                             'label' => 'group',
-                            'class' => 'App:Labels',
+                            'class' => Labels::class,
                             'query_builder' => function (LabelRepository $er) use ($roomId) {
                                 return $er->createQueryBuilder('l')
                                     ->andWhere('l.contextId = :contextId')
@@ -69,7 +68,7 @@ class RubricFilterType extends AbstractType
                     $builder
                         ->add('topic', Filters\EntityFilterType::class, array(
                             'label' => 'topic',
-                            'class' => 'App:Labels',
+                            'class' => Labels::class,
                             'query_builder' => function (LabelRepository $er) use ($roomId) {
                                 return $er->createQueryBuilder('l')
                                     ->andWhere('l.contextId = :contextId')
@@ -84,12 +83,12 @@ class RubricFilterType extends AbstractType
                         ))
                     ;
                 }
-                
+
                 // institution
                 if (in_array('institution', $filterableRubrics)) {
                     $builder
                         ->add('institution', Filters\EntityFilterType::class, array(
-                            'class' => 'App:Labels',
+                            'class' => Labels::class,
                             'query_builder' => function (LabelRepository $er) use ($roomId) {
                                 return $er->createQueryBuilder('l')
                                     ->andWhere('l.contextId = :contextId')
@@ -112,7 +111,7 @@ class RubricFilterType extends AbstractType
      * Returns the prefix of the template block name for this type.
      * The block prefix defaults to the underscored short class name with the "Type" suffix removed
      * (e.g. "UserProfileType" => "user_profile").
-     * 
+     *
      * @return string The prefix of the template block name
      */
     public function getBlockPrefix()
